test(EuiDataGridCell): move EuiDataGrid keyboard controls jest tests to a separate file
